requirements.txt
Flask>=3.0.3
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Flask-WTF>=1.2.1
SQLAlchemy>=2.0.36
Werkzeug>=3.0.6
python-dotenv>=1.0.1
WTForms>=3.1.2
